requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
jsonpath-python==1.0.6
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index-core==0.12.23.post2
llama-index-readers-file==0.4.6
marshmallow==3.26.1
mistralai==1.5.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.65.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
poppler-utils==0.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.1
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
